requirements.txt
flask==3.0.0
flask-cors==4.0.0
chromadb==0.4.22
langchain==0.1.0
langchain-community==0.0.10
pypdf==3.17.4
python-docx==1.1.0
sentence-transformers==2.2.2
ollama==0.1.6
werkzeug==3.0.1